Add kernel address sanitizer infrastructure.
This patch only adds infrastructure for kernel address sanitizer. It's not
available for use yet. The idea (and a bit of code as well) was borrowed
from [1]. Thanks to Dmitriy Vyukov and Andrey Konovalov.

Kernel address sanitizer (KASAN) is a dynamic memory error detector. It provides
fast and comprehensive solution for finding use-after-free and out-of-bounds bugs.
It also possible to find unitialized memory accesses, but it's not implemented yet.

The main features of KASAN is:
 - is based on compiler instrumentation (fast),
 - detects out of bounds for both writes and reads,
 - provides use after free detection,
 - does prompt detection of bad memory accesses,
 - prints informative reports.

The main idea of KASAN use shadow memory to record whether each byte of memory
is safe to access, and use instrumentation to check the shadow memory on each
memory access. To check every memory access KASAN requires the kernel to be built
with a specific compiler. Before each memory accesse of size 1, 2, 4, 8 or 16
 GCC inserts a specific function calls (__kasan_read1, _kasan_write1, ...,
__kasan_read16, __kasan_write16). These functions check if the accessed memory
region is valid to access and print a error if it's not.

[1] https://code.google.com/p/address-sanitizer/wiki/AddressSanitizerForKernel

Signed-off-by: Andrey Ryabinin <a.ryabinin@samsung.com>